Don't extract test names from output, since it runs at eval time
We want individual derivations for each test, so they're easier to track,
etc. (e.g. in Hydra).

We were parsing these out of the test results, but that requires having the
results at eval time, which in turn means running the tests at eval time, and in
turn means building all of the packages (ML4PG, Emacs, etc.) at eval time.

This is pretty naff, so we now run a quick grep/awk combo over the Emacs Lisp
files in ./test to extract the names. This can be done at eval time without
requiring any other stuff to be built or run. Unfortunately there's a chance
that the test names we infer using this approach differ from those which
actually get run. To spot such things, we include a couple of sanity checks:

 - One runs at eval time, to ensure the list of 'test suites' being run ("coq"
   and "ssreflect") matches the test suites we found when inferring names.
 - The other runs during the build of every test, and checks that all test names
   we inferred for that test suite correspond to those in the results of that
   suite. Since this uses the results, we defer it to build time. Since tests
   might be missing, we can't just check if *this* test was found; we must check
   that all names match up.